+++ test_cli.py
@@ -1,9 +1,99 @@
 import unittest
 import sys
+from io import StringIO
 from cli import play, show_board_with_icons
 from unittest.mock import patch, MagicMock
 
 class TestCli(unittest.TestCase):
+
+    @patch('sys.stdout', new_callable=StringIO)
+    def test_show_board_with_icons(self, mock_stdout):
+
+        '''
+        La función test_show_board_with_icons es una prueba unitaria que verifica si el tablero de ajedrez se muestra correctamente con los iconos de las piezas.
+        La función test_show_board_with_icons verifica que, cuando se llama a la función show_board_with_icons con un tablero de ajedrez como argumento, 
+        el tablero se imprime correctamente con los iconos de las piezas. El parámetro mock_stdout es un objeto simulado que reemplaza la salida estándar durante la prueba.
+        '''
+        board = [
+            ['R', 'N', 'B', 'Q', 'K', 'B', 'N', 'R'],
+            ['P', 'P', 'P', 'P', 'P', 'P', 'P', 'P'],
+            ['.', '.', '.', '.', '.', '.', '.', '.'],
+            ['.', '.', '.', '.', '.', '.', '.', '.'],
+            ['.', '.', '.', '.', '.', '.', '.', '.'],
+            ['.', '.', '.', '.', '.', '.', '.', '.'],
+            ['p', 'p', 'p', 'p', 'p', 'p', 'p', 'p'],
+            ['r', 'n', 'b', 'q', 'k', 'b', 'n', 'r']
+        ]
+        show_board_with_icons(board)
+        expected_output = (
+            '♜ ♞ ♝ ♛ ♚ ♝ ♞ ♜\n'
+            '♟ ♟ ♟ ♟ ♟ ♟ ♟ ♟\n'
+            '· · · · · · · ·\n'
+            '· · · · · · · ·\n'
+            '· · · · · · · ·\n'
+            '· · · · · · · ·\n'
+            '♙ ♙ ♙ ♙ ♙ ♙ ♙ ♙\n'
+            '♖ ♘ ♗ ♕ ♔ ♗ ♘ ♖\n'
+        )
+        self.assertEqual(mock_stdout.getvalue(), expected_output)
+
+    @patch('builtins.input', side_effect=['SALIR'])
+    @patch('builtins.print')
+    def test_salir_from_row(self, mock_print, mock_input):
+
+        '''
+        La función test_salir es una prueba unitaria que verifica cómo se maneja la entrada "SALIR" en el tablero de ajedrez.
+        La función test_salir verifica que, cuando se simula la entrada del usuario con el valor "SALIR" en lugar de la cooerdenada de la fila origen, 
+        el programa maneja la entrada correctamente sin intentar mover ninguna pieza, muestra el mensaje "Juego terminado." y finaliza el juego. 
+        '''
+        chess = MagicMock()
+        with self.assertRaises(SystemExit):
+            play(chess)
+        mock_print.assert_called_with("Juego terminado.")
+
+    @patch('builtins.input', side_effect=['1', 'SALIR'])
+    @patch('builtins.print')
+    def test_salir_from_col(self, mock_print, mock_input):
+
+        '''
+        La función test_salir es una prueba unitaria que verifica cómo se maneja la entrada "SALIR" en el tablero de ajedrez.
+        La función test_salir verifica que, cuando se simula la entrada del usuario con el valor "SALIR" en lugar de la cooerdenada de la columna origen,
+        el programa maneja la entrada correctamente sin intentar mover ninguna pieza, muestra el mensaje "Juego terminado." y finaliza el juego.
+        '''
+        chess = MagicMock()
+        with self.assertRaises(SystemExit):
+            play(chess)
+        mock_print.assert_called_with("Juego terminado.")
+
+    @patch('builtins.input', side_effect=['1', '1', 'SALIR'])
+    @patch('builtins.print')
+    def test_salir_to_row(self, mock_print, mock_input):
+
+        '''
+        La función test_salir es una prueba unitaria que verifica cómo se maneja la entrada "SALIR" en el tablero de ajedrez.
+        La función test_salir verifica que, cuando se simula la entrada del usuario con el valor "SALIR" en lugar de la cooerdenada de la fila destino,
+        el programa maneja la entrada correctamente sin intentar mover ninguna pieza, muestra el mensaje "Juego terminado." y finaliza el juego.
+        '''
+        chess = MagicMock()
+        with self.assertRaises(SystemExit):
+            play(chess)
+        mock_print.assert_called_with("Juego terminado.")
+
+    @patch('builtins.input', side_effect=['1', '1', '2', 'SALIR'])
+    @patch('builtins.print')
+    def test_salir_to_col(self, mock_print, mock_input):
+
+        '''
+        La función test_salir es una prueba unitaria que verifica cómo se maneja la entrada "SALIR" en el tablero de ajedrez.
+        La función test_salir verifica que, cuando se simula la entrada del usuario con el valor "SALIR" en lugar de la cooerdenada de la columna destino,
+        el programa maneja la entrada correctamente sin intentar mover ninguna pieza, muestra el mensaje "Juego terminado." y finaliza el juego.
+        '''
+        chess = MagicMock()
+        with self.assertRaises(SystemExit):
+            play(chess)
+        mock_print.assert_called_with("Juego terminado.")
+
+
     @patch('builtins.input', side_effect=['6', '0', '4', '0'])
     @patch('cli.show_board_with_icons')
     def test_movimiento_valido_peon(self, mock_show_board, mock_input):
@@ -55,7 +145,6 @@ class TestCli(unittest.TestCase):
         chess = MagicMock()
         chess.get_board.return_value = [['.']*8 for _ in range(8)]
         chess.turn = 'white'
-        from io import StringIO
         captured_output = StringIO()
         sys.stdout = captured_output
         play(chess)
@@ -68,7 +157,6 @@ class TestCli(unittest.TestCase):
         chess = MagicMock()
         chess.get_board.return_value = [['.']*8 for _ in range(8)]
         chess.turn = 'white'
-        from io import StringIO
         captured_output = StringIO()
         sys.stdout = captured_output
         play(chess)
@@ -81,7 +169,6 @@ class TestCli(unittest.TestCase):
         chess = MagicMock()
         chess.get_board.return_value = [['.']*8 for _ in range(8)]
         chess.turn = 'white'
-        from io import StringIO
         captured_output = StringIO()
         sys.stdout = captured_output
         play(chess)
@@ -94,7 +181,6 @@ class TestCli(unittest.TestCase):
         chess = MagicMock()
         chess.get_board.return_value = [['.']*8 for _ in range(8)]
         chess.turn = 'white'
-        from io import StringIO
         captured_output = StringIO()
         sys.stdout = captured_output
         play(chess)
@@ -117,7 +203,6 @@ class TestCli(unittest.TestCase):
         chess = MagicMock()
         chess.get_board.return_value = [['.']*8 for _ in range(8)]
         chess.turn = 'white'
-        from io import StringIO
         captured_output = StringIO()
         sys.stdout = captured_output
         play(chess)
@@ -130,7 +215,6 @@ class TestCli(unittest.TestCase):
         chess = MagicMock()
         chess.get_board.return_value = [['.']*8 for _ in range(8)]
         chess.turn = 'white'
-        from io import StringIO
         captured_output = StringIO()
         sys.stdout = captured_output
         play(chess)
@@ -143,7 +227,6 @@ class TestCli(unittest.TestCase):
         chess = MagicMock()
         chess.get_board.return_value = [['.']*8 for _ in range(8)]
         chess.turn = 'white'
-        from io import StringIO
         captured_output = StringIO()
         sys.stdout = captured_output
         play(chess)
@@ -156,7 +239,6 @@ class TestCli(unittest.TestCase):
         chess = MagicMock()
         chess.get_board.return_value = [['.']*8 for _ in range(8)]
         chess.turn = 'white'
-        from io import StringIO
         captured_output = StringIO()
         sys.stdout = captured_output
         play(chess)