+++ CHANGELOG.md
@@ -1,17 +1,26 @@
 # Changelog
 
-## [0.0.3] - 20-08-2024
+## [0.4] - 20-08-2024
+
+### Agregado:
+-Creacion de los test de las piezas.
+
+### Mejoras
+-Mejoras del test de la interfaz de usuario.
+-Mejoras de los movimientos posibles del peon.
+
+## [0.3] - 20-08-2024
 
 ### Agregado:
 -Creacion de los test de la interfaz de usuario.
 -Salir de juego con el string SALIR
 
-## [0.0.2] - 15-08-2024
+## [0.2] - 15-08-2024
 
 ### Agregado:
 -Creacion de los movimientos de las piezas.
 
-## [0.0.1] - 13-08-2024
+## [0.1] - 13-08-2024
 
 ### Agregado:
 -Creación de la interfaz.